- divided into 2 classes BeanConfiguration.java
